requirements.txt
Django==5.1.1
django-stubs==5.0.4
python-decouple==3.8
psycopg2==2.9.10
channels==4.1.0
daphne==4.1.2
django-allauth==65.2.0
requests==2.32.3
pyjwt==2.9.0
types-requests==2.32.0.20241016
uvicorn==0.32.1
websockets==13.1
wsproto==1.2.0
selenium==4.26.1